+++ src/main/java/com/github/nkzawa/socketio/client/Manager.java
@@ -1,5 +1,6 @@
 package com.github.nkzawa.socketio.client;
 
+import com.github.nkzawa.backo.Backoff;
 import com.github.nkzawa.emitter.Emitter;
 import com.github.nkzawa.socketio.parser.Packet;
 import com.github.nkzawa.socketio.parser.Parser;
@@ -77,13 +78,13 @@ public class Manager extends Emitter {
     private boolean reconnecting;
     private boolean reconnectFailed; /* added by matt */
     private boolean encoding;
-    private boolean openReconnect;
     private int _reconnectionAttempts;
     private long _reconnectionDelay;
     private long _reconnectionDelayMax;
+    private double _randomizationFactor;
+    private Backoff backoff;
     private long _timeout;
     private Set<Socket> connected;
-    private int attempts;
     private URI uri;
     private List<Packet> packetBuffer;
     private Queue<On.Handle> subs;
@@ -96,9 +97,6 @@ public class Manager extends Emitter {
      * This HashMap can be accessed from outside of EventThread.
      */
     private ConcurrentHashMap<String, Socket> nsps;
-
-    private ScheduledExecutorService timeoutScheduler;
-    private ScheduledExecutorService reconnectScheduler;
 
 
     public Manager() {
@@ -130,11 +128,15 @@ public class Manager extends Emitter {
         this.reconnectionAttempts(opts.reconnectionAttempts != 0 ? opts.reconnectionAttempts : Integer.MAX_VALUE);
         this.reconnectionDelay(opts.reconnectionDelay != 0 ? opts.reconnectionDelay : 1000);
         this.reconnectionDelayMax(opts.reconnectionDelayMax != 0 ? opts.reconnectionDelayMax : 5000);
+        this.randomizationFactor(opts.randomizationFactor != 0.0 ? opts.randomizationFactor : 0.5);
+        this.backoff = new Backoff()
+                .setMin(this.reconnectionDelay())
+                .setMax(this.reconnectionDelayMax())
+                .setJitter(this.randomizationFactor());
         this.timeout(opts.timeout < 0 ? 20000 : opts.timeout);
         this.readyState = ReadyState.CLOSED;
         this.uri = uri;
         this.connected = new HashSet<Socket>();
-        this.attempts = 0;
         this.encoding = false;
         this.packetBuffer = new ArrayList<Packet>();
         this.encoder = new Parser.Encoder();
@@ -145,6 +147,15 @@ public class Manager extends Emitter {
         this.emit(event, args);
         for (Socket socket : this.nsps.values()) {
             socket.emit(event, args);
+        }
+    }
+
+    /**
+     * Update `socket.id` of all sockets
+     */
+    private void updateSocketIds() {
+        for (Socket socket : this.nsps.values()) {
+            socket.id = this.engine.id();
         }
     }
 
@@ -172,6 +183,21 @@ public class Manager extends Emitter {
 
     public Manager reconnectionDelay(long v) {
         this._reconnectionDelay = v;
+        if (this.backoff != null) {
+            this.backoff.setMin(v);
+        }
+        return this;
+    }
+
+    public double randomizationFactor() {
+        return this._randomizationFactor;
+    }
+
+    public Manager randomizationFactor(double v) {
+        this._randomizationFactor = v;
+        if (this.backoff != null) {
+            this.backoff.setJitter(v);
+        }
         return this;
     }
 
@@ -181,6 +207,9 @@ public class Manager extends Emitter {
 
     public Manager reconnectionDelayMax(long v) {
         this._reconnectionDelayMax = v;
+        if (this.backoff != null) {
+            this.backoff.setMax(v);
+        }
         return this;
     }
 
@@ -195,8 +224,7 @@ public class Manager extends Emitter {
 
     private void maybeReconnectOnOpen() {
         // Only try to reconnect if it's the first time we're connecting
-        if (!this.openReconnect && !this.reconnecting && this._reconnection) {
-            this.openReconnect = true;
+        if (!this.reconnecting && this._reconnection && this.backoff.getAttempts() == 0) {
             this.reconnect();
         }
     }
@@ -253,9 +281,10 @@ public class Manager extends Emitter {
                             Exception err = new SocketIOException("Connection error",
                                     data instanceof Exception ? (Exception) data : null);
                             fn.call(err);
+                        } else {
+                            // Only do this if there is no fn to handle the error
+                            self.maybeReconnectOnOpen();
                         }
-
-                        self.maybeReconnectOnOpen();
                     }
                 });
 
@@ -263,7 +292,8 @@ public class Manager extends Emitter {
                     final long timeout = Manager.this._timeout;
                     logger.fine(String.format("connection attempt will timeout after %d", timeout));
 
-                    final Future timer = getTimeoutScheduler().schedule(new Runnable() {
+                    final Timer timer = new Timer();
+                    timer.schedule(new TimerTask() {
                         @Override
                         public void run() {
                             EventThread.exec(new Runnable() {
@@ -277,12 +307,12 @@ public class Manager extends Emitter {
                                 }
                             });
                         }
-                    }, timeout, TimeUnit.MILLISECONDS);
+                    }, timeout);
 
                     Manager.this.subs.add(new On.Handle() {
                         @Override
                         public void destroy() {
-                            timer.cancel(false);
+                            timer.cancel();
                         }
                     });
                 }
@@ -375,6 +405,7 @@ public class Manager extends Emitter {
                 socket.on(Socket.EVENT_CONNECT, new Listener() {
                     @Override
                     public void call(Object... objects) {
+                        s.id = self.engine.id();
                         self.connected.add(s);
                     }
                 });
@@ -429,7 +460,11 @@ public class Manager extends Emitter {
     }
 
     /*package*/ void close() {
+        if (this.readyState != ReadyState.OPEN) {
+            this.cleanup();
+        }
         this.skipReconnect = true;
+        this.backoff.reset();
         this.readyState = ReadyState.CLOSED;
         if (this.engine != null) {
             this.engine.close();
@@ -439,15 +474,9 @@ public class Manager extends Emitter {
     private void onclose(String reason) {
         logger.fine("close");
         this.cleanup();
+        this.backoff.reset();
         this.readyState = ReadyState.CLOSED;
         this.emit(EVENT_CLOSE, reason);
-
-        if (this.timeoutScheduler != null) {
-            this.timeoutScheduler.shutdown();
-        }
-        if (this.reconnectScheduler != null) {
-            this.reconnectScheduler.shutdown();
-        }
 
         if (this._reconnection && !this.skipReconnect) {
             this.reconnect();
@@ -458,21 +487,20 @@ public class Manager extends Emitter {
         if (this.reconnecting || reconnectFailed /* added by matt*/ || this.skipReconnect) return;
 
         final Manager self = this;
-        this.attempts++;
 
-        if (attempts > this._reconnectionAttempts) {
+        if (this.backoff.getAttempts() >= this._reconnectionAttempts) {
             logger.fine("reconnect failed");
+            this.backoff.reset();
             this.emitAll(EVENT_RECONNECT_FAILED);
-            this.attempts = 0; /* added by matt*/
             this.reconnectFailed = true; /* added by matt*/
             this.reconnecting = false;
         } else {
-            long delay = this.attempts * this.reconnectionDelay();
-            delay = Math.min(delay, this.reconnectionDelayMax());
+            long delay = this.backoff.duration();
             logger.fine(String.format("will wait %dms before reconnect attempt", delay));
 
             this.reconnecting = true;
-            final Future timer = this.getReconnectScheduler().schedule(new Runnable() {
+            final Timer timer = new Timer();
+            timer.schedule(new TimerTask() {
                 @Override
                 public void run() {
                     EventThread.exec(new Runnable() {
@@ -481,8 +509,9 @@ public class Manager extends Emitter {
                             if (self.skipReconnect) return;
 
                             logger.fine("attempting reconnect");
-                            self.emitAll(EVENT_RECONNECT_ATTEMPT, self.attempts);
-                            self.emitAll(EVENT_RECONNECTING, self.attempts);
+                            int attempts = self.backoff.getAttempts();
+                            self.emitAll(EVENT_RECONNECT_ATTEMPT, attempts);
+                            self.emitAll(EVENT_RECONNECTING, attempts);
 
                             // check again for the case socket closed in above events
                             if (self.skipReconnect) return;
@@ -504,36 +533,23 @@ public class Manager extends Emitter {
                         }
                     });
                 }
-            }, delay, TimeUnit.MILLISECONDS);
+            }, delay);
 
             this.subs.add(new On.Handle() {
                 @Override
                 public void destroy() {
-                    timer.cancel(false);
+                    timer.cancel();
                 }
             });
         }
     }
 
     private void onreconnect() {
-        int attempts = this.attempts;
-        this.attempts = 0;
+        int attempts = this.backoff.getAttempts();
         this.reconnecting = false;
+        this.backoff.reset();
+        this.updateSocketIds();
         this.emitAll(EVENT_RECONNECT, attempts);
-    }
-
-    private ScheduledExecutorService getTimeoutScheduler() {
-        if (this.timeoutScheduler == null || this.timeoutScheduler.isShutdown()) {
-           this.timeoutScheduler = Executors.newSingleThreadScheduledExecutor();
-        }
-        return timeoutScheduler;
-    }
-
-    private ScheduledExecutorService getReconnectScheduler() {
-        if (this.reconnectScheduler == null || this.reconnectScheduler.isShutdown()) {
-            this.reconnectScheduler = Executors.newSingleThreadScheduledExecutor();
-        }
-        return this.reconnectScheduler;
     }
 
 
@@ -556,6 +572,7 @@ public class Manager extends Emitter {
         public int reconnectionAttempts;
         public long reconnectionDelay;
         public long reconnectionDelayMax;
+        public double randomizationFactor;
         public long timeout = -1;
     }
 }