+++ src/test/java/com/github/nkzawa/socketio/client/Connection.java
@@ -7,9 +7,14 @@ import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
 import java.net.URISyntaxException;
+import java.util.HashMap;
+import java.util.Map;
 import java.util.concurrent.*;
+import java.util.logging.Logger;
 
 public abstract class Connection {
+
+    private static final Logger logger = Logger.getLogger(Connection.class.getName());
 
     final static int TIMEOUT = 7000;
     final static int PORT = 3000;
@@ -21,7 +26,7 @@ public abstract class Connection {
 
     @Before
     public void startServer() throws IOException, InterruptedException {
-        System.out.println("Starting server ...");
+        logger.fine("Starting server ...");
 
         final CountDownLatch latch = new CountDownLatch(1);
         serverProcess = Runtime.getRuntime().exec(
@@ -37,10 +42,10 @@ public abstract class Connection {
                     line = reader.readLine();
                     latch.countDown();
                     do {
-                        System.out.println("SERVER OUT: " + line);
+                        logger.fine("SERVER OUT: " + line);
                     } while ((line = reader.readLine()) != null);
                 } catch (IOException e) {
-                    e.printStackTrace();
+                    logger.warning(e.getMessage());
                 }
             }
         });
@@ -52,10 +57,10 @@ public abstract class Connection {
                 String line;
                 try {
                     while ((line = reader.readLine()) != null) {
-                        System.err.println("SERVER ERR: " + line);
+                        logger.fine("SERVER ERR: " + line);
                     }
                 } catch (IOException e) {
-                    e.printStackTrace();
+                    logger.warning(e.getMessage());
                 }
             }
         });
@@ -64,7 +69,7 @@ public abstract class Connection {
 
     @After
     public void stopServer() throws InterruptedException {
-        System.out.println("Stopping server ...");
+        logger.fine("Stopping server ...");
         serverProcess.destroy();
         serverOutput.cancel(false);
         serverError.cancel(false);
@@ -95,6 +100,16 @@ public abstract class Connection {
     }
 
     String[] createEnv() {
-        return new String[] {"DEBUG=socket.io:*", "PORT=" + PORT};
+        Map<String, String> env = new HashMap<String, String>(System.getenv());
+        env.put("DEBUG", "socket.io:*");
+        env.put("PORT", String.valueOf(PORT));
+        String[] _env = new String[env.size()];
+        int i = 0;
+        for (String key : env.keySet()) {
+            _env[i] = key + "=" + env.get(key);
+            i++;
+        }
+        return _env;
+
     }
 }